Extract vanity, add restore
