+++ subkey/Cargo.toml
@@ -6,6 +6,7 @@ authors = ["Parity Technologies <admin@parity.io>"]
 [dependencies]
 substrate-primitives = { version = "*", path = "../core/primitives" }
 rand = "0.4"
+clap = { version = "~2.32", features = ["yaml"] }
 
 [features]
 bench = []
